Updated Tracker to something that builds again.
I truly hate Django after these last couple of days.
I hope it will burn in the fires of dark net for all eternity.

Thanks PUWP for the django-paypal fixes to make it work on a still available and somewhat modern version.
Thanks Zoton2 for directing me to said fixes when I was on the verge of throwing the whole project off a cliff, and maybe myself of another cliff.
Because who would want to share a grave with this mess?

Oh, and AMQP integration for donations. Finally after only what, two years?
Yay!